♻️ refactor : separate context in popover component, add displayname
