+++ packages/popover/src/components/Popover.tsx
@@ -1,7 +1,7 @@
 /** @jsxImportSource @emotion/react */
 import { filterComponent } from '@jdesignlab/react-utils';
 import { createClassVariant } from '@jdesignlab/theme';
-import { PopoverProvider } from './PopoverContext';
+import { PopoverProvider } from './PopoverProvider';
 import { PopoverHeader } from './PopoverHeader';
 import { PopoverTrigger } from './PopoverTrigger';
 import { PopoverBody } from './PopoverBody';
@@ -10,12 +10,12 @@ import { PopoverContent } from './PopoverContent';
 import type { PopoverProps } from '../types';
 
 export const Popover = (props: PopoverProps) => {
-  const { children, ...propsWithoutChildren } = props;
+  const { children, ...restProps } = props;
   const Trigger = filterComponent(children, Popover.Trigger, true);
   const PopoverContents = filterComponent(children, Popover.Trigger, false);
 
   return (
-    <PopoverProvider popoverProps={propsWithoutChildren}>
+    <PopoverProvider popoverProps={restProps}>
       <div className={createClassVariant('popover', 'wrapper')} css={{ position: 'relative', display: 'inline' }}>
         {Trigger}
         {PopoverContents}
@@ -24,6 +24,7 @@ export const Popover = (props: PopoverProps) => {
   );
 };
 
+Popover.displayName = 'Popover';
 Popover.Trigger = PopoverTrigger;
 Popover.Content = PopoverContent;
 Popover.Header = PopoverHeader;
